Fix consistency of CIDR JSON example
I first thought the field name for CIDR conditions always was ``remoteIpAddress``. When reading the tests code, I realize the field name in context had to match the condition key.

I wonder whether it's necessary to phrase it explicitly, now that I know it :) But at least this tiny fix will help a bit!

Thanks for the awesome library!